third_party/WebKit/LayoutTests/ambient-light/create-event.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/js-test.js"></script>
<script>
description("Tests the DeviceLightEvent constructor.");

var newEvent = new DeviceLightEvent("devicelight", {
    bubbles: true, cancelable: false,
    value: 10
});

var defaultEvent = new DeviceLightEvent("devicelight");

shouldBeTrue("typeof newEvent.type == 'string'");
shouldBeEqualToString("newEvent.type", "devicelight");
shouldBeTrue("typeof newEvent.bubbles == 'boolean'");
shouldBeTrue("newEvent.bubbles");
shouldBeTrue("typeof newEvent.cancelable == 'boolean'");
shouldBeFalse("newEvent.cancelable")
shouldBeTrue("typeof newEvent.value == 'number'");
shouldBeEqualToNumber('newEvent.value', 10);

// FIXME: Consider making bubbles property configurable.
shouldBeTrue("defaultEvent.bubbles");
shouldBeFalse("defaultEvent.cancelable");
shouldBe("defaultEvent.value", "Infinity");
</script>
</body>
</html>
