third_party/WebKit/LayoutTests/ambient-light/create-event-expected.txt

Tests the DeviceLightEvent constructor.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS typeof newEvent.type == 'string' is true
PASS newEvent.type is "devicelight"
PASS typeof newEvent.bubbles == 'boolean' is true
PASS newEvent.bubbles is true
PASS typeof newEvent.cancelable == 'boolean' is true
PASS newEvent.cancelable is false
PASS typeof newEvent.value == 'number' is true
PASS newEvent.value is 10
PASS defaultEvent.bubbles is true
PASS defaultEvent.cancelable is false
PASS defaultEvent.value is Infinity
PASS successfullyParsed is true

TEST COMPLETE
